requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
argparse==1.4.0
    # via -r requirements.in
attrs==24.2.0
    # via
    #   service-identity
    #   twisted
automat==24.8.1
    # via twisted
azure-core==1.32.0
    # via azure-storage-blob
azure-storage-blob==12.24.0
    # via -r requirements.in
blinker==1.9.0
    # via flask
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via flask
colorama==0.4.6
    # via click
constantly==23.10.4
    # via twisted
contourpy==1.3.1
    # via matplotlib
cryptography==44.0.0
    # via
    #   azure-storage-blob
    #   pyopenssl
    #   scrapy
    #   service-identity
cssselect==1.2.0
    # via
    #   parsel
    #   scrapy
cycler==0.12.1
    # via matplotlib
defusedxml==0.7.1
    # via scrapy
dnspython==2.7.0
    # via pymongo
filelock==3.16.1
    # via tldextract
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==5.0.0
    # via -r requirements.in
fonttools==4.55.0
    # via matplotlib
gpxpy==1.6.2
    # via -r requirements.in
hyperlink==21.0.0
    # via twisted
idna==3.10
    # via
    #   hyperlink
    #   requests
    #   tldextract
incremental==24.7.2
    # via twisted
isodate==0.7.2
    # via azure-storage-blob
itemadapter==0.10.0
    # via
    #   itemloaders
    #   scrapy
itemloaders==1.3.2
    # via scrapy
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
jmespath==1.0.1
    # via
    #   itemloaders
    #   parsel
joblib==1.4.2
    # via scikit-learn
kiwisolver==1.4.7
    # via matplotlib
lxml==5.3.0
    # via
    #   parsel
    #   scrapy
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.9.3
    # via seaborn
numpy==2.1.3
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
packaging==24.2
    # via
    #   matplotlib
    #   parsel
    #   scrapy
pandas==2.2.3
    # via
    #   -r requirements.in
    #   seaborn
parsel==1.9.1
    # via
    #   itemloaders
    #   scrapy
pillow==11.0.0
    # via matplotlib
protego==0.3.1
    # via scrapy
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.1
    # via service-identity
pycparser==2.22
    # via cffi
pydispatcher==2.0.7
    # via scrapy
pymongo==4.10.1
    # via -r requirements.in
pyopenssl==24.3.0
    # via scrapy
pyparsing==3.2.0
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2024.2
    # via pandas
queuelib==1.7.0
    # via scrapy
requests==2.32.3
    # via
    #   azure-core
    #   requests-file
    #   tldextract
requests-file==2.1.0
    # via tldextract
scikit-learn==1.5.2
    # via -r requirements.in
scipy==1.14.1
    # via scikit-learn
scrapy==2.12.0
    # via -r requirements.in
seaborn==0.13.2
    # via -r requirements.in
service-identity==24.2.0
    # via scrapy
six==1.16.0
    # via
    #   azure-core
    #   python-dateutil
threadpoolctl==3.5.0
    # via scikit-learn
tldextract==5.1.3
    # via scrapy
twisted==24.11.0
    # via scrapy
typing-extensions==4.12.2
    # via
    #   azure-core
    #   azure-storage-blob
    #   twisted
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
w3lib==2.2.1
    # via
    #   parsel
    #   scrapy
werkzeug==3.1.3
    # via flask
zope-interface==7.2
    # via
    #   scrapy
    #   twisted
python-dotenv==0.19.1
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
